refactor: seperate codes into packages
