+++ nextstep-di/src/main/java/nextstep/di/factory/factory/BeanScanner.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.factory;
 
 import com.google.common.collect.Sets;
 import nextstep.annotation.ComponentScan;
@@ -12,8 +12,6 @@ import java.util.Set;
 public class BeanScanner {
     private static final Logger log = LoggerFactory.getLogger(BeanScanner.class);
     private Reflections reflections;
-    private Class<ComponentScan> componentScanClass = ComponentScan.class;
-    private Set<String> basePackages;
 
     public BeanScanner(Object... basePackages) {
         reflections = new Reflections(basePackages);
