+++ nextstep-di/src/test/java/nextstep/di/factory/BeanFactoryTest.java
@@ -1,7 +1,16 @@
 package nextstep.di.factory;
 
 import nextstep.annotation.Configuration;
-import nextstep.di.factory.example.*;
+import nextstep.di.factory.example.controller.QnaController;
+import nextstep.di.factory.example.controller.TestController;
+import nextstep.di.factory.example.repository.JdbcQuestionRepository;
+import nextstep.di.factory.example.repository.QuestionRepository;
+import nextstep.di.factory.example.repository.TestJdbcTemplate;
+import nextstep.di.factory.example.service.MyQnaService;
+import nextstep.di.factory.example.service.TestService;
+import nextstep.di.factory.factory.BasePackageFinder;
+import nextstep.di.factory.factory.BeanFactory;
+import nextstep.di.factory.factory.BeanScanner;
 import nextstep.stereotype.Controller;
 import nextstep.stereotype.Repository;
 import nextstep.stereotype.Service;
