+++ nextstep-di/src/test/java/nextstep/di/factory/example/configuration/TestJdbcConfiguration.java
@@ -1,15 +1,17 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.configuration;
 
 import nextstep.annotation.Bean;
 import nextstep.annotation.ComponentScan;
 import nextstep.annotation.Configuration;
+import nextstep.di.factory.example.repository.TestJdbcTemplate;
+import nextstep.di.factory.example.service.TestService;
 import org.apache.commons.dbcp2.BasicDataSource;
 
 import javax.sql.DataSource;
 
 @Configuration
-@ComponentScan(basePackages = {"nextstep.annotation"})
-public class ExampleConfig {
+@ComponentScan(basePackages = {"nextstep.di.factory.example"})
+public class TestJdbcConfiguration {
     @Bean
     public DataSource dataSource() {
         BasicDataSource ds = new BasicDataSource();
@@ -18,5 +20,10 @@ public class ExampleConfig {
         ds.setUsername("sa");
         ds.setPassword("");
         return ds;
+    }
+
+    @Bean
+    public TestJdbcTemplate jdbcTemplate(DataSource dataSource, TestService testService) {
+        return new TestJdbcTemplate(dataSource, testService);
     }
 }
