+++ nextstep-di/src/test/java/nextstep/di/factory/example/repository/QuestionRepository.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.repository;
 
 import nextstep.stereotype.Repository;
