+++ nextstep-di/src/test/java/nextstep/di/factory/example/repository/TestUserDao.java
@@ -1,0 +1,4 @@
+package nextstep.di.factory.example.repository;
+
+public class TestUserDao {
+}
