+++ nextstep-di/src/test/java/nextstep/di/factory/example/service/MyQnaService.java
@@ -1,13 +1,17 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.service;
 
+import nextstep.annotation.Inject;
+import nextstep.di.factory.example.repository.QuestionRepository;
+import nextstep.di.factory.example.repository.UserRepository;
 import nextstep.stereotype.Service;
 
 @Service
-public class TestService {
+public class MyQnaService {
     private UserRepository userRepository;
     private QuestionRepository questionRepository;
 
-    public TestService(UserRepository userRepository, QuestionRepository questionRepository) {
+    @Inject
+    public MyQnaService(UserRepository userRepository, QuestionRepository questionRepository) {
         this.userRepository = userRepository;
         this.questionRepository = questionRepository;
     }
